modularizing [categories, tools, platforms]
