+++ src/db/general.ts
@@ -9,25 +9,6 @@ import {
 // --- Core Tables ---
 
 
-// --- Categories ---
-
-export const categories = pgTable('categories', {
-  id: uuid('id').primaryKey().defaultRandom(),
-  name: text('name').notNull(),
-  imageUrl: text('image_url'),
-  description: text('description'),
-  json: jsonb('json'),
-});
-
-// --- Platforms ---
-
-export const platforms = pgTable('platforms', {
-  id: uuid('id').primaryKey().defaultRandom(),
-  name: text('name').notNull(),
-  imageUrl: text('image_url'),
-  description: text('description'),
-  json: jsonb('json'),
-});
 
 // --- Tags ---
 
@@ -64,15 +45,7 @@ export const companies = pgTable('companies', {
 
 // --- Main Entity ---
 
-// --- Tools ---
 
-export const tools = pgTable('tools', {
-  id: uuid('id').primaryKey().defaultRandom(),
-  name: text('name').notNull(),
-  imageUrl: text('image_url'),
-  description: text('description'),
-  json: jsonb('json'),
-});
 
 // --- Join Tables ---
