+++ src/db/schema.ts
@@ -1,3 +1,6 @@
 import { account, session, user, verification } from "./auth-schema"
+import { tools } from "./tools";
+import { categories } from "./categories";
+import { platforms } from "./platforms";
 
-export { account, session, user, verification }
+export { account, session, user, verification, tools, categories, platforms }
